requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
font-fredoka-one==0.0.4
font-hanken-grotesk==0.0.2
font-intuitive==0.0.4
font-source-serif-pro==0.0.1
future==1.0.0
geocoder==1.38.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
importlib-metadata==6.7.0
inky==1.5.0
numpy==1.21.6
openai==1.37.0
Pillow==9.5.0
pkg-resources==0.0.0
pydantic==2.5.3
pydantic-core==2.14.6
ratelim==0.1.6
requests==2.31.0
RPi.GPIO==0.7.1
six==1.16.0
smbus2==0.4.3
sniffio==1.3.1
soupsieve==2.4.1
spidev==3.6
tqdm==4.66.4
typing-extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
